chore: Update dependencies

+++ modules/config/config.go
@@ -66,7 +66,7 @@ var (
 	AllowedEmailsRe = `^[a-zA-Z0-9.!#$%&’*+/=?^_\x60{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)+$`
 	AllowedImagesRe = `^https:\/\/(www\.)?(userstyles.world/api|(user-images|raw)\.githubusercontent\.com|github\.com|gitlab\.com|codeberg\.org|cdn\.jsdelivr\.net/gh/33kk)\/.*\.(jpe?g|png|webp)(\?inline=(true|false))?$`
 
-	CachedCodeItems = getEnvInt("CACHED_CODE_ITEMS", 25)
+	CachedCodeItems = uint(getEnvInt("CACHED_CODE_ITEMS", 25))
 )
 
 // OAuthURL returns the proper callback URL depending on the environment.
